Render basic list of prices
Renders a minimal unstyled list of assets along with their icons and
prices. This currently only fetches the data once (polling to be
added).

Sets up Redux boilerplate for managing the prices state, in this case
a list of currencies and their metadata/price along with whether the
data has been fetched.